Addresing review comments
