+++ src/apify/storages/__init__.py
@@ -1,5 +1,5 @@
 from crawlee.storages import Dataset, KeyValueStore, RequestQueue
 
-from .request_list import RequestList
+from ._request_list import RequestList
 
 __all__ = ['Dataset', 'KeyValueStore', 'RequestQueue', 'RequestList']
